Add: ✨ 'need' on running windows command
We added 'need' to run which job first that needs to finish before it
can execute the running windows command. This solution is to make sure
that we don't run our job parallel, but rather "steps" you can say.
So if Step A is not working, then we can't run Step B so to say.