Add: ✨ Greet input/output from github action from a repository
Github action takes someone codes, or your code that does some specific
task, and that you can use in your steps in your GitHub workflow.
This task can be simple as logging something, but of course it can
be usedin some complicated task like maybe deploying a Web Site
or something.

By using actions, we can write less code in our workflows and also
we can use a code that somebody else shared that does some common task
or we can actually create our own actions and share it with people
so they can use it in their own workflow.

The github action takes reference from one of this:

- commit SHA ID
- github version tag
- branch

It is not recommended to use branches, for example main-branch.
This is because breaking changes most likely will happen on the main
branch, and therefore whatever types of github actions we are trying
to do, it will most likely fail and break.

It is better to use primarly github version tag. Secondary, you could
use github commit SHA ID

You can read more about the github action documentation steps

Source:
- https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idsteps
- https://github.com/actions/hello-world-javascript-action