Add: ✨ Filtering Workflows by Branches, Tags & Paths
We can tell the github action by using which branches, tags and files
to do certain task when of our workflow. This can be very effective
if we want to do some testing for example.

We can filter what we want to include/exclude which files we don't
need it to run testing as well.

Make sure to read this documentation here to understanad the pattern
which files to matches if we want to include/exclude file/branches/paths:

Source:
https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#filter-pattern-cheat-sheet